+++ projects/demo/src/app/app.config.ts
@@ -4,13 +4,13 @@ import {
   importProvidersFrom,
 } from '@angular/core';
 import { provideRouter } from '@angular/router';
+import { NgxZenConfig, NgxZenModule } from 'ngx-zen';
 
 import { routes } from './app.routes';
-import { NgxZenConfig, NgxZenModule, ZenTheme } from 'ngx-zen';
 
 const ngxConfig: NgxZenConfig = {
-  theme: 'dark',
-  shape: 'curve',
+  shape: 'pill',
+  btnSize: 'wide',
 };
 
 export const appConfig: ApplicationConfig = {