+++ projects/ngx-zen/button/button.component.ts
@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { Component, Input } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
 
 @Component({
   selector: 'ngx-zen-button',
@@ -8,11 +8,11 @@ import { Component, Input } from '@angular/core';
   templateUrl: './button.component.html',
   styleUrls: ['./button.component.scss'],
 })
-export class NgxButtonComponent {
+export class NgxButtonComponent implements OnInit {
   @Input() text = 'Submit';
   @Input() zenClass: string = '';
-  @Input() shape: 'pill' | 'curve' | 'default' = 'default';
-  @Input() size: 'compact' | 'wide' | 'tight' | 'default' = 'default';
+  @Input() shape!: 'pill' | 'curve' | 'default';
+  @Input() size!: 'compact' | 'wide' | 'tight' | 'default';
   @Input() type: 'icononly' | 'default' = 'default';
   @Input() icon!: string;
   @Input() iconPosition: 'left' | 'right' = 'left';
@@ -20,4 +20,17 @@ export class NgxButtonComponent {
   @Input() imgPosition: 'left' | 'right' = 'left';
   @Input() variant: 'outlined' | 'default' | 'link' = 'default';
   @Input() disabled: boolean = false;
+
+  private globalConfig: { shape: string, btnSize: string } = { shape: '', btnSize: '' };
+
+  ngOnInit(): void {
+    const rootStyles = getComputedStyle(document.documentElement);
+    this.globalConfig.shape = rootStyles.getPropertyValue('--shape').trim();
+    this.globalConfig.btnSize = rootStyles.getPropertyValue('--btn-size').trim();
+  }
+  
+  get buttonClasses(): string[] {
+    return [this.type, this.shape || this.globalConfig.shape, this.size || this.globalConfig.btnSize, this.variant, this.zenClass]
+      .filter(cls => cls && cls !== 'default');
+  }
 }