+++ projects/ngx-zen/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Beta version 0.0.40 Updates
+
+- Root config format changed, see the root config documentation for full
+- Curve shape changed from 4px to 6px
+
 ## Beta version 0.0.39 Updates
 
 - Badge variant types updated to be consistent